requirements.txt
torch>=1.9.0
torchvision>=0.10.0
networkx>=2.6
matplotlib>=3.3.0
numpy>=1.21.0
flask>=2.0.0
flask-cors>=3.0.0
pillow>=8.0.0
markupsafe>=2.0.0
requests>=2.25.0



